+++ app/src/main/java/com/mortarnav/MainActivity.java
@@ -1,24 +1,27 @@
 package com.mortarnav;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 
-import com.mortarnav.deps.WithAppDependencies;
+import com.mortarnav.deps.WithActivityDependencies;
 import com.mortarnav.stackable.HomeStackable;
 import com.mortarnav.view.MyPopup2View;
 import com.mortarnav.view.MyPopupView;
+
+import javax.inject.Inject;
 
 import architect.Navigator;
 import architect.NavigatorView;
 import architect.TransitionsMapping;
 import architect.commons.ActivityArchitector;
 import architect.commons.Architected;
+import architect.commons.transition.BottomAppearTransition;
+import architect.commons.transition.Config;
+import architect.commons.transition.FadeModalTransition;
+import architect.commons.transition.LateralViewTransition;
 import architect.robot.DaggerService;
-import architect.transition.BottomAppearTransition;
-import architect.transition.Config;
-import architect.transition.FadeModalTransition;
-import architect.transition.LateralViewTransition;
 import autodagger.AutoComponent;
 import autodagger.AutoInjector;
 import butterknife.Bind;
@@ -33,17 +36,23 @@ import mortar.bundler.BundleServiceRunner;
  */
 @AutoComponent(
         dependencies = App.class,
-        superinterfaces = WithAppDependencies.class
+        superinterfaces = WithActivityDependencies.class
 )
 @AutoInjector
 @DaggerScope(MainActivity.class)
-public class MainActivity extends Activity {
+public class MainActivity extends AppCompatActivity {
 
     private MortarScope scope;
     private Navigator navigator;
 
+    @Inject
+    protected ToolbarOwner toolbarOwner;
+
     @Bind(R.id.navigator_container)
     protected NavigatorView containerView;
+
+    @Bind(R.id.toolbar)
+    protected Toolbar toolbar;
 
     @Override
     public Object getSystemService(String name) {
@@ -58,14 +67,22 @@ public class MainActivity extends Activity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
+        setContentView(R.layout.activity_main);
+        ButterKnife.bind(this);
+
+        toolbar.setTitle("Mortar architect");
+        setSupportActionBar(toolbar);
+
         scope = ActivityArchitector.onCreateScope(this, savedInstanceState, new Architected() {
             @Override
             public Navigator createNavigator(MortarScope scope) {
                 Navigator navigator = Navigator.create(scope, new Parceler());
                 navigator.transitions().register(new TransitionsMapping()
-                        .byDefault(new LateralViewTransition(new Config().duration(300)))
-                        .show(MyPopupView.class).withTransition(new FadeModalTransition(new Config().duration(250)))
-                        .show(MyPopup2View.class).withTransition(new BottomAppearTransition(false, new Config().duration(1000))));
+                                .byDefault(new LateralViewTransition(new Config().duration(300)))
+                                .show(MyPopupView.class).withTransition(new FadeModalTransition(new Config().duration(250)))
+                                .show(MyPopup2View.class).withTransition(new BottomAppearTransition(true, new Config().duration(1000)))
+//                                .show(SlidesView.class).withTransition(new CustomFullScreenLateralTransition())
+                );
                 return navigator;
             }
 
@@ -74,15 +91,14 @@ public class MainActivity extends Activity {
                 MainActivityComponent component = DaggerMainActivityComponent.builder()
                         .appComponent(parentScope.<AppComponent>getService(DaggerService.SERVICE_NAME))
                         .build();
-                component.inject(MainActivity.this);
-
                 builder.withService(DaggerService.SERVICE_NAME, component);
             }
         });
 
-        setContentView(R.layout.activity_main);
-        ButterKnife.bind(this);
+        DaggerService.<MainActivityComponent>get(this).inject(this);
+        toolbarOwner.takeView(toolbar);
 
+        // it is usually the best to create the navigator after everything else
         navigator = ActivityArchitector.onCreateNavigator(this, savedInstanceState, containerView, new HomeStackable("Default home path"));
     }
 
@@ -113,6 +129,8 @@ public class MainActivity extends Activity {
 
     @Override
     protected void onDestroy() {
+        toolbarOwner.dropView(toolbar);
+
         navigator.delegate().onDestroy();
         navigator = null;