+++ app/src/main/java/com/mortarnav/view/HomeView.java
@@ -1,23 +1,34 @@
 package com.mortarnav.view;
 
+import android.animation.Animator;
+import android.animation.AnimatorListenerAdapter;
+import android.animation.AnimatorSet;
 import android.content.Context;
 import android.view.View;
 import android.widget.TextView;
 
 import com.mortarnav.R;
+import com.mortarnav.ToolbarOwner;
 import com.mortarnav.presenter.HomePresenter;
 import com.mortarnav.stackable.HomeStackable;
 
+import javax.inject.Inject;
+
 import architect.commons.view.PresentedScrollView;
 import architect.robot.DaggerService;
+import architect.view.HandlesViewTransition;
 import butterknife.Bind;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
+import timber.log.Timber;
 
 /**
  * @author Lukasz Piliszczuk - lukasz.pili@gmail.com
  */
-public class HomeView extends PresentedScrollView<HomePresenter> {
+public class HomeView extends PresentedScrollView<HomePresenter> implements HandlesViewTransition {
+
+    @Inject
+    protected ToolbarOwner toolbarOwner;
 
     @Bind(R.id.home_title)
     public TextView titleTextView;
@@ -83,5 +94,32 @@ public class HomeView extends PresentedScrollView<HomePresenter> {
     @OnClick(R.id.home_set_new_stack)
     void setNewStackClick() {
         presenter.setNewStackClick();
+    }
+
+
+    @Override
+    public void onViewTransition(AnimatorSet set) {
+        if (set != null) {
+            set.addListener(new AnimatorListenerAdapter() {
+
+                @Override
+                public void onAnimationStart(Animator animation) {
+                    toolbarOwner.show();
+                }
+
+                @Override
+                public void onAnimationEnd(Animator animation) {
+                    toolbarOwner.setTitle("Hello Home!");
+                }
+            });
+            Animator animator = toolbarOwner.animateShow();
+            Timber.d("Animate show: %s", animator);
+            if (animator != null) {
+                set.play(animator);
+            }
+        } else {
+            toolbarOwner.show();
+            toolbarOwner.setTitle("Hello Home!");
+        }
     }
 }