+++ app/src/main/java/com/mortarnav/view/SubnavView.java
@@ -1,16 +1,23 @@
 package com.mortarnav.view;
 
+import android.animation.Animator;
+import android.animation.AnimatorListenerAdapter;
+import android.animation.AnimatorSet;
 import android.content.Context;
 import android.view.View;
 
 import com.mortarnav.R;
+import com.mortarnav.ToolbarOwner;
 import com.mortarnav.presenter.SubnavPresenter;
 import com.mortarnav.presenter.stackable.SubnavStackableComponent;
+
+import javax.inject.Inject;
 
 import architect.NavigatorView;
 import architect.commons.view.PresentedFrameLayout;
 import architect.robot.DaggerService;
 import architect.view.HandlesBack;
+import architect.view.HandlesViewTransition;
 import autodagger.AutoInjector;
 import butterknife.Bind;
 import butterknife.ButterKnife;
@@ -19,7 +26,10 @@ import butterknife.ButterKnife;
  * @author Lukasz Piliszczuk - lukasz.pili@gmail.com
  */
 @AutoInjector(SubnavPresenter.class)
-public class SubnavView extends PresentedFrameLayout<SubnavPresenter> implements HandlesBack {
+public class SubnavView extends PresentedFrameLayout<SubnavPresenter> implements HandlesBack, HandlesViewTransition {
+
+    @Inject
+    protected ToolbarOwner toolbarOwner;
 
     @Bind(R.id.sub_navigator)
     public NavigatorView navigatorView;
@@ -36,5 +46,20 @@ public class SubnavView extends PresentedFrameLayout<SubnavPresenter> implements
     @Override
     public boolean onBackPressed() {
         return presenter.backPressed();
+    }
+
+
+    @Override
+    public void onViewTransition(AnimatorSet set) {
+        if (set != null) {
+            set.addListener(new AnimatorListenerAdapter() {
+                @Override
+                public void onAnimationEnd(Animator animation) {
+                    toolbarOwner.setTitle("Subnav presenter!");
+                }
+            });
+        } else {
+            toolbarOwner.setTitle("Subnav presenter!");
+        }
     }
 }