+++ architect/src/main/java/architect/TransitionsMapping.java
@@ -5,8 +5,6 @@ import android.view.View;
 import java.util.ArrayList;
 import java.util.List;
 
-import architect.transition.ViewTransition;
-
 /**
  * Views transitions mapping
  *