+++ architect/src/main/java/architect/ViewTransitionDirection.java
@@ -3,6 +3,6 @@ package architect;
 /**
  * @author Lukasz Piliszczuk - lukasz.pili@gmail.com
  */
-public enum TransitionDirection {
-    FORWARD, BACKWARD
+public enum ViewTransitionDirection {
+    FORWARD, BACKWARD, REPLACE
 }