+++ commons/src/main/java/architect/commons/transition/FadeModalTransition.java
@@ -1,13 +1,15 @@
-package architect.transition;
+package architect.commons.transition;
 
 import android.animation.AnimatorSet;
 import android.animation.ObjectAnimator;
 import android.view.View;
 
+import architect.ViewTransitionDirection;
+
 /**
  * @author Lukasz Piliszczuk - lukasz.pili@gmail.com
  */
-public class FadeModalTransition extends BaseModalTransition<View> {
+public class FadeModalTransition extends ModalTransition {
 
     protected boolean hideExitView;
 
@@ -34,13 +36,14 @@ public class FadeModalTransition extends BaseModalTransition<View> {
     }
 
     @Override
-    public void show(View view, AnimatorSet set) {
-        set.play(ObjectAnimator.ofFloat(view, View.ALPHA, 0, 1));
-    }
+    public void transition(View enterView, View exitView, ViewTransitionDirection direction, AnimatorSet set) {
+        super.transition(enterView, exitView, direction, set);
 
-    @Override
-    public void hide(View view, AnimatorSet set) {
-        set.play(ObjectAnimator.ofFloat(view, View.ALPHA, 1, 0));
+        if (direction == ViewTransitionDirection.FORWARD) {
+            set.play(ObjectAnimator.ofFloat(enterView, View.ALPHA, 0, 1));
+        } else {
+            set.play(ObjectAnimator.ofFloat(exitView, View.ALPHA, 1, 0));
+        }
     }
 
     @Override